Replace custom rho with existing exp. value calculation
